fix(psalm phase1): null-safety and typing fixes across controllers, entities, repositories, and Jira service\n\n- project: normalize internalJiraTicketSystem to nullable\n- admin: guard nulls on delete; validate preset relations before save\n- crud: ensure ticket prefix not null; guard internal jira ticketsystem; validate issue object\n- entry: strict ints for DateTime::setDate; cast rounded duration to int\n- interpretation: safe parse_str on nullable query string\n- repositories: tighten return types and guard array sorts\n- entry-repo: null guards for customer/project/activity/ticket\n- jira: minor tidy for readability

+++ src/Entity/Entry.php
@@ -398,7 +398,7 @@ class Entry extends Base
         if (!$start instanceof DateTime) {
             $start = new DateTime($start);
             [$year, $month, $day] = explode('-', $this->getDay()->format('Y-m-d'));
-            $start->setDate($year, $month, $day);
+            $start->setDate((int) $year, (int) $month, (int) $day);
         }
 
         $this->start = $start;
@@ -426,7 +426,7 @@ class Entry extends Base
         if (!$end instanceof DateTime) {
             $end = new DateTime($end);
             [$year, $month, $day] = explode('-', $this->getDay()->format('Y-m-d'));
-            $end->setDate($year, $month, $day);
+            $end->setDate((int) $year, (int) $month, (int) $day);
         }
 
         $this->end = $end;
@@ -623,7 +623,7 @@ class Entry extends Base
             $end = new DateTime($this->getEnd()->format('H:i'));
 
             $difference = ($end->getTimestamp() - $start->getTimestamp()) * $factor / 60;
-            $this->setDuration(round($difference));
+            $this->setDuration((int) round($difference));
         } else {
             $this->setDuration(0);
         }
